requirements.txt
aioimaplib
rich
websockets
playwright=1.44.0
playwright-stealth=1.0.6
python-dotenv